+++ benchmark/matrix_blas/matmul.cpp
@@ -5,7 +5,7 @@
 #define CRS_PERF 2 * M *N *nnzrow / time / 1.0e+9
 
 template <typename MAT_A, typename MAT_B, typename MAT_C, typename T>
-bool benchmark(const size_t M, const size_t N, const size_t K, int iter) {
+bool benchmark(const size_t M, const size_t N, const size_t K, size_t iter) {
 
   size_t nnzrow = 81;
   if ((nnzrow < M) && (nnzrow < N) && (nnzrow < K)) {
@@ -71,7 +71,7 @@ int main(int argc, char **argv) {
 
   std::cout << "func\tprec\tM\tN\tK\ttime[sec]\tperf[GFLOPS] " << std::endl;
 
-  int iter = MATRIX_BENCH_ITER;
+  size_t iter = MATRIX_BENCH_ITER;
 
   // monolish::util::set_log_level(3);
   // monolish::util::set_log_filename("./monolish_log.txt");
